Extract Pubgrub resolver into separate files
